// frontend/lib/schema.ts
export interface components {
  schemas: {
    Dataproduct: {
      id: string
      name: string
      description?: string
      slug: string
      repo?: string
      last_modified: string
      created: string
      owner: components['schemas']['Owner']
      keyword?: string[]
      datasets?: {
        id?: string
        type?: components['schemas']['DatasetType']
      }[]
    }
    DatasetType: 'bigquery'
    Owner: {
      team: string
      teamkatalogen?: string
    }
    NewDataproduct: {
      name: string
      description?: string
      slug?: string
      repo?: string
      owner: components['schemas']['Owner']
      keyword?: string[]
    }
    Dataset: {
      id?: string
      dataproduct_id?: string
      name?: string
      description?: string
      pii?: boolean
      bigquery?: {
        project_id: string
        dataset: string
        table: string
      }
    }
    NewDataset: {
      name: string
      description?: string
      pii: boolean
      bigquery: {
        project_id: string
        dataset: string
        table: string
      }
    }
    SearchResultEntry: {
      url?: string
      type?: 'dataset' | 'dataproduct' | 'datapackage'
      id?: string
      name?: string
      excerpt?: string
    }
  }
}